% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizability.R
\name{alpha_bare}
\alias{alpha_bare}
\title{alpha_bare}
\usage{
alpha_bare(wavelength = seq(300, 800), alpha_inf = 9.6e-39,
  alpha_k = c(5.76e-38), lambda_k = c(526), mu_k = c(10000))
}
\arguments{
\item{wavelength}{wavelength in nm}

\item{alpha_inf}{scalar real offset}

\item{alpha_k}{vector of oscillator strengths}

\item{lambda_k}{vector of oscillator wavelengths}

\item{mu_k}{vector of oscillator damping terms}
}
\value{
data.frame
}
\description{
Bare (intrinsic) polarizability of a dye in vacuum
}
\details{
Sum of lorentz oscillators
}
\author{
baptiste Auguie
}
\seealso{
Other user_level polarisability: \code{\link{alpha_dye}},
  \code{\link{alpha_ellipsoid}}
}

