% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizability.R
\name{alpha_dye}
\alias{alpha_dye}
\title{alpha_dye}
\usage{
alpha_dye(sizes, wavelength, medium, ...)
}
\arguments{
\item{sizes}{matrix of particle sizes (scaling factors for polarisability tensor)}

\item{wavelength}{wavelength in nm}

\item{medium}{refractive index of incident medium}

\item{...}{further parameters passed to the Lorentzian function}
}
\value{
matrix of polarisability
}
\description{
Principal polarisability components for a dye molecule
}
\details{
The dye is modelled as a sum of Lorentz oscillators
}
\author{
baptiste Auguie
}
\seealso{
Other user_level polarisability: \code{\link{alpha_bare}},
  \code{\link{alpha_ellipsoid}}
}

