\name{array}
\alias{array}
\title{
	Rcpp module: array
}
\description{
	Exposes a C++ calculation of the array factor. 
}
\details{
	The module contains the following items: 
	
	functions: \describe{
        \item{array_factor}{ truncated lattice sum for a finite 2D square array}
		} 
	
	
}

\examples{
show( array )
}
\keyword{datasets}
