% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cda-package.r
\docType{data}
\name{array}
\alias{array}
\title{Rcpp module: array}
\description{
Exposes a C++ calculation of the array factor.
}
\details{
\itemize{
 \item{array_factor}{ Truncated lattice sum for a finite 2D square array}
}
}
\examples{
show( array )
}

