\name{cd}
\alias{cd}
\title{
	Rcpp module: cd
}
\description{
	Exposes a calculation of orientation-averaged circular dichroism within the coupled-dipole approximation.
}
\details{
	The module contains the following items: 
	
	functions: \describe{
        \item{circular_dichroism_spectrum}{Loop over wavelenghts and calculate the orientation averaging of the difference in extinction, absorption, scattering for left/right circularly polarised light.}
		} 
	
	
}
\examples{
show( cd )
}
\keyword{datasets}
