% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda-package.r
\docType{package}
\name{cda-package}
\alias{cda-package}
\title{cda}
\description{
Coupled-dipole approximation for electromagnetic scattering by three-dimensional clusters of subwavelength particles
}
\details{
Coupled-dipole simulations for electromagnetic scattering of light by subwavelength 
particles in arbitrary 3-dimensional configurations. Scattering and absorption 
spectra simulated by direct or iterative inversion of the interaction matrix. 
Functions are provided for linear polarisation with varying angle of incidence, 
as well as circular polarisation with angular averaging (optical activity).
}
\author{
baptiste Auguie
}
\references{
Draine BT. The discrete-dipole approximation and its application to interstellar graphite grains. Astrophysical Journal. 1988.
## Any one of the following references should be used to cite and acknowledge the use of this package.

Circular dichroism:

B. Auguie, J.L. Alonso-Gomez, A. Guerrero-Martinez, L.M. Liz-Marzan. Fingers crossed: circular dichroism with a dimer of plasmonic nanorods. J. Phys. Chem. Lett. 2, (2011)

Linear extinction:

B. Auguie, W.L. Barnes. Diffractive coupling in gold nanoparticle arrays and the effect of disorder. Optics Letters (2009)

Array factor (infinite case):

B. Auguie, W.L. Barnes. Collective resonances in gold nanoparticle arrays. Physical Review Letters (2008)
}
\keyword{package}
\keyword{packagelibrary}

