\name{clust.equalsizes}
\alias{clust.equalsizes}
\title{clust.equalsizes}
\usage{clust.equalsizes(a, b, c, N)
}
\description{clust.equalsizes}
\details{generate a matrix of equal particle sizes}
\value{matrix Nx3}
\author{baptiste Auguie}
\arguments{\item{a}{a}
\item{b}{b}
\item{c}{c}
\item{N}{N}
}

