% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_ball}
\alias{cluster_ball}
\title{cluster_ball}
\usage{
cluster_ball(N, R0 = 15, a = 1, b = 1, c = b)
}
\arguments{
\item{N}{number of particles}

\item{R0}{ball radius}

\item{a}{semi axis}

\item{b}{semi axis}

\item{c}{semi axis}
}
\value{
list of class cluster with fields: positions, sizes,  angles
}
\description{
cluster_ball
}
\details{
a ball of dipoles on a cubic lattice
}
\examples{
b = cluster_ball(1000); 
library(rgl)
open3d()
rgl.spheres(t(b$positions))
rgl.spheres(0,0, 0, radius = 15, alpha=0.2, col="red")
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_array}},
  \code{\link{cluster_chain}}, \code{\link{cluster_dimer}},
  \code{\link{cluster_helix}}, \code{\link{cluster_shell}},
  \code{\link{cluster_single}}
}

