\name{cluster_chain}
\alias{cluster_chain}
\title{cluster_chain}
\usage{
  cluster_chain(N, pitch = 500, a = 50, b = 30, c = b)
}
\arguments{
  \item{N}{number of rods}

  \item{pitch}{pitch}

  \item{a}{semi axis}

  \item{b}{semi axis}

  \item{c}{semi axis}
}
\value{
  list with r, sizes, angles
}
\description{
  cluster_chain
}
\details{
  linear chain of parallel nanorods
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level cluster: \code{\link{cluster_dimer}},
  \code{\link{cluster_helix}}, \code{\link{equal_angles}},
  \code{\link{equal_sizes}}, \code{\link{helix}}
}

