% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_shell}
\alias{cluster_shell}
\title{cluster_shell}
\usage{
cluster_shell(N = 50, R0 = 30, d = 1, a = 1, b = 1, c = 1,
  orientation = c("radial", "flat", "random"), position = c("fibonacci",
  "hc", "random"), exclusion = 5 * N^(-1/4), seed = 123, ...)
}
\arguments{
\item{N}{number of particles}

\item{R0}{radius of core}

\item{d}{distance}

\item{a}{ellipsoid semi-axis}

\item{b}{ellipsoid semi-axis}

\item{c}{ellipsoid semi-axis}

\item{position}{type of random coverage}

\item{seed}{random seed for reproducibility}

\item{...}{extra arguments (ignored)}

\item{position}{type of angular orientation}
}
\value{
list
}
\description{
cluster_shell
}
\details{
shell cluster
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_array}},
  \code{\link{cluster_ball}}, \code{\link{cluster_chain}},
  \code{\link{cluster_dimer}}, \code{\link{cluster_helix}},
  \code{\link{cluster_single}}
}

