% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_single}
\alias{cluster_single}
\title{cluster_single}
\usage{
cluster_single(a, b = a, c = b, phi = 0, theta = 0, psi = 0)
}
\arguments{
\item{a}{semi axis}

\item{b}{semi axis}

\item{c}{semi axis}

\item{phi}{angle}

\item{theta}{angle}

\item{psi}{angle}
}
\value{
list of class cluster with fields: positions, sizes,  angles
}
\description{
cluster_single
}
\details{
a single particle
}
\examples{
b = cluster_ball(1000); 
library(rgl)
open3d()
rgl.spheres(t(b$positions))
rgl.spheres(0,0, 0, radius = 15, alpha=0.2, col="red")
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_array}},
  \code{\link{cluster_ball}}, \code{\link{cluster_chain}},
  \code{\link{cluster_dimer}}, \code{\link{cluster_helix}},
  \code{\link{cluster_shell}}
}

