% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizability.R
\name{depolarisation}
\alias{depolarisation}
\title{depolarisation}
\usage{
depolarisation(x1, x2 = x1, x3 = x2)
}
\arguments{
\item{x1}{semi-axis in nm}

\item{x2}{semi-axis in nm}

\item{x3}{semi-axis in nm}
}
\value{
shape factor along x1
}
\description{
Depolarisation factor for an ellipsoid
}
\details{
calculates the 3 depolarisation factors for a general ellipsoid
}
\author{
baptiste Auguie
}
\seealso{
Other user_level polarizability: \code{\link{alpha_kuwata}}
}

