% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{equal_angles}
\alias{equal_angles}
\title{equal_angles}
\usage{
equal_angles(phi, theta, gamma, N)
}
\arguments{
\item{phi}{Euler angle}

\item{theta}{Euler angle}

\item{gamma}{Euler angle}

\item{N}{number of particles}
}
\value{
3xN matrix
}
\description{
Utility function to create clusters
}
\details{
Identical particles
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda utility: \code{\link{dye_coverage}},
  \code{\link{equal_sizes}}, \code{\link{spheroid_ar}}
}

