% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{equal_sizes}
\alias{equal_sizes}
\title{equal_sizes}
\usage{
equal_sizes(a, b, c, N)
}
\arguments{
\item{a}{semi-axis along x}

\item{b}{semi-axis along y}

\item{c}{semi-axis along z}

\item{N}{number of particles}
}
\value{
3xN matrix
}
\description{
Utility function to create clusters
}
\details{
Identical particles
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda utility: \code{\link{dye_coverage}},
  \code{\link{equal_angles}}, \code{\link{spheroid_ar}}
}

