\name{exercise_14.15}
\docType{data}
\alias{exercise_14.15}
\title{Data for Exercise 14.15 on Buchanan Vote in 2000}
\description{Data for Exercise 14.15 on Buchanan vote in 2000.}
\usage{data(exercise_14.15)}
\format{A data frame containing 67 observations (rows)
  and the following columns:
  \describe{
  \item{county}{}
  \item{perot_vote}{}
  \item{total_vote_1996}{}
  \item{buchanan_vote}{}
  \item{total_vote_2000}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
