\name{exercise_4.13}
\docType{data}
\alias{exercise_4.13}
\title{Ray Allen Data Set for Exercise 4.13}
\description{Ray Allen data set for Exercise 4.13.}
\usage{data(exercise_4.13)}
\format{A data frame containing 24 observations (rows)
  and the following columns:
  \describe{
  \item{game}{}
  \item{made}{}
  \item{attempts}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
