\name{exercise_9.5}
\docType{data}
\alias{exercise_9.5}
\title{Government Spending Data Set for Exercise 9.5}
\description{Government spending data set for Exercise 9.5.}
\usage{data(exercise_9.5)}
\format{A data frame containing 81 observations (rows)
  and the following columns:
  \describe{
  \item{e}{}
  \item{h}{}
  \item{c}{}
  \item{l}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
