% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{data}
\name{G0}
\alias{G0}
\title{Precomputed array factor for a square lattice at normal incidence}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{wavelength}{normalised wavelength lambda/pitch}
  \item{Qx}{in-plane component of the wavevector (0, since normal incidence)}
  \item{Gxx}{complex value of the array factor}
}}
\source{
Javier Garcia de Abajo
}
\usage{
G0
}
\description{
Exact calculation of the array factor using code from Javier Garcia de Abajo 
(part of the pxtal program for multiple scattering calculations 
in infinite layered 2D arrays)
}
\keyword{datasets}

