% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_directions.R
\name{getCanonicalDirections}
\alias{getCanonicalDirections}
\title{Calculate canonical parallel and orthogonal directions}
\usage{
getCanonicalDirections(X, Y, COD = TRUE, method = "comp")
}
\arguments{
\item{X}{a p-feature by n-sample \code{matrix}}

\item{Y}{a p-feature by n-sample \code{matrix},
features and samples must align with X}

\item{COD}{Specifies whether to calculate the COD or not}

\item{method}{'eigen' uses slow R matrix multiplication to get the
covariance matrix. 'comp' uses the considerably faster prcomp method
from irlba (most users will want this option)}
}
\value{
a list containing:
\item{directions}{p by 1 vector of canonical directions}
\item{projections}{n by 1 vectors of projections of the data
  onto the canonical directions. Example: proj_vcpd_X is the
  projection of X onto the CPD direction}
}
\description{
Implements the canonical orthogonal and parallel directions
from Chapter 2 of Xuxin Liu's 2007 dissertation
}
