% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{data}
\name{gfun}
\alias{gfun}
\title{Precomputed array factor for a square lattice at normal incidence}
\format{A list of two interpolation functions:
\describe{
  \item{re}{real part of G0}
  \item{im}{imaginary part of G0}
}}
\source{
Javier Garcia de Abajo
}
\usage{
gfun
}
\description{
Exact calculation of the array factor using code from Javier Garcia de Abajo 
(part of the pxtal program for multiple scattering calculations 
in infinite layered 2D arrays)
}
\keyword{datasets}

