% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{helix}
\alias{helix}
\title{helix}
\usage{
helix(R0 = 500, pitch = 600, N = 5, delta = pi/8, delta0 = pi/2,
  n.smooth = 100 * N, right = TRUE)
}
\arguments{
\item{R0}{radius of helix}

\item{pitch}{pitch of helix}

\item{N}{number of particles}

\item{delta}{angle between particles}

\item{delta0}{initial angle}

\item{n.smooth}{number of points for a finer helix (useful for display)}

\item{right}{logical, helicity}
}
\value{
list of positions and angles
}
\description{
Positions along a helix
}
\details{
3D points following a helix
}
\author{
baptiste Auguie
}

