\name{linear}
\alias{linear}
\title{
	Rcpp module: linear
}
\description{
	Exposes C++ calculation of scattering and absorption of dipolar particles by linearly polarised light in fixed orientation.
}
\details{
	The module contains the following items: 
	
	functions: \describe{
        \item{linear_extinction_spectrum}{ Loop over wavelengths, calculate the extinction and scattering for linearly polarised light in two orthogonal directions. }
		} 
	
	
}
\examples{
show( linear )
}
\keyword{datasets}
