\docType{data}
\name{linear}
\alias{linear}
\title{Rcpp module: linear}
\description{
  Exposes C++ calculation of scattering and absorption of
  dipolar particles by linearly polarised light in fixed
  orientation.
}
\details{
  \itemize{ \item{linear_extinction_spectrum}{Returns
  extinction spectra for x and y polarisation at one fixed
  incidence } \item{dispersion}{Returns absorption and
  extinction cross-sections for x and y polarisation at
  multiple angles of incidence, fixed wavelength
  (subroutine)} \item{dispersion_spectrum}{Returns
  absorption and extinction spectra for x and y
  polarisation at multiple angles of incidence.} }
}
\examples{
show( linear )
}
\keyword{datasets}

