\name{linear_extinction_spectrum}
\alias{linear_extinction_spectrum}
\title{extinction_spectrum}
\usage{linear_extinction_spectrum(cluster, material, n=1.33, progress=FALSE)
}
\description{extinction spectrum}
\details{extinction spectrum}
\value{data.frame}
\author{baptiste Auguie}
\arguments{\item{cluster}{list describing a cluster}
\item{material}{list}
\item{n}{medium refractive index}
\item{progress}{logical, display progress bar}
}

