\name{makeHelixCluster}
\alias{makeHelixCluster}
\title{makeHelixCluster}
\usage{makeHelixCluster(N=5, R0=0.012, pitch=0.015, delta=pi/2, delta0=0, right=TRUE, a=0.005,
    b=a, c=a, angles=c("helix", "random", "fixed"), seed=123, ...)
}
\description{makeHelixCluster}
\details{makeHelixCluster}
\value{list}
\author{baptiste Auguie}
\arguments{\item{N}{number of particles}
\item{R0}{radius of helix}
\item{pitch}{pitch of helix}
\item{delta}{angle between particles}
\item{delta0}{initial angle}
\item{right}{logical, helicity}
\item{a}{ellipsoid semi-axis}
\item{b}{ellipsoid semi-axis}
\item{c}{ellipsoid semi-axis}
\item{angles}{type of angular orientation}
\item{seed}{random seed for reproducibility}
\item{...}{ignored arguments}
}

