\name{makeRodChain}
\alias{makeRodChain}
\title{makeRodChain}
\usage{
  makeRodChain(N, pitch = 0.5, a = 0.05, b = 0.03, c = b)
}
\arguments{
  \item{N}{number of rods}

  \item{pitch}{pitch}

  \item{a}{semi axis}

  \item{b}{semi axis}

  \item{c}{semi axis}
}
\value{
  list with r, sizes, angles
}
\description{
  makeRodChain
}
\details{
  makeRodChain
}
\author{
  baptiste Auguie
}

