\name{particles.povray}
\alias{particles.povray}
\title{particles.povray}
\usage{particles.povray(positions, angles, sizes, out="positions.pov")
}
\description{create particles for povray input}
\details{writes a list of particles}
\value{side-effect only}
\author{baptiste Auguie}
\arguments{\item{positions}{matrix of positions}
\item{angles}{matrix of Euler angles in radians}
\item{sizes}{matrix of particle sizes}
\item{out}{output filename}
}

