% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.r
\name{quadrature_sphere}
\alias{quadrature_sphere}
\title{quadrature_sphere}
\usage{
quadrature_sphere(Nq = 30, quadrature = c("qmc", "gl", "cheap", "random"),
  init = TRUE)
}
\arguments{
\item{Nq}{number of integration points}

\item{quadrature}{quadrature method, using either Gauss Legendre quadrature (default), Quasi Monte Carlo, regular grid, or "cheap" (3 axes)}

\item{init}{(qmc method only) logical: restart, or continue from previous call}
}
\description{
Quadrature points on a sphere
}
\details{
Numerical integration points for angular averaging
}
\author{
baptiste Auguie
}

