% Generated by roxygen2 (4.0.0): do not edit by hand

\name{rgl_annotate}
\alias{rgl_annotate}
\title{rgl_annotate}
\usage{
rgl_annotate()
}
\value{
draw axes
}
\description{
Add axes to a rgl scene
}
\details{
x, y, z axes
}
\author{
baptiste Auguie
}
\seealso{
Other user_level rgl: \code{\link{rgl.ellipsoid}},
\code{\link{rgl.ellipsoids}}
}

