% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sample.R
\name{sample_random}
\alias{sample_fibonacci}
\alias{sample_hc}
\alias{sample_random}
\title{Generate a random sample of points on the unit sphere}
\usage{
sample_random(N)

sample_hc(N, exclusion = 0.1, maxiter = 200L, k = 30L)

sample_fibonacci(N = 301)
}
\arguments{
\item{N}{number of points}

\item{exclusion}{minimum distance allowed between points}

\item{maxiter}{maximum number of iterations}

\item{k}{number of extra new points to try at each iteration}
}
\value{
3xN matrix

3xN matrix
}
\description{
Random sample

Random sample with minimum exlusion zone ("hard-core process")

Fibonacci coverage of a sphere
}
\details{
Produces a set of points that covers rather uniformly the unit sphere with N points
with a spiral-like pattern based on a Fibonacci sequence
}
\section{Functions}{
\itemize{
\item \code{sample_random}: random sample

\item \code{sample_hc}: random sample with exclusion zone

\item \code{sample_fibonacci}: 
}}
\examples{
sample_random(10)
sample_hc(10)
}
\author{
baptiste Auguie
}

