% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{spectrum_dispersion}
\alias{spectrum_dispersion}
\title{spectrum_dispersion}
\usage{
spectrum_dispersion(cluster, material, medium = 1.33, Incidence = 0,
  Axes = "z", polarisation = c("linear", "circular"), method = c("ls",
  "oos"), cg = FALSE, born = FALSE, Nsca = 50, maxiter = 30,
  tol = 1e-04, progress = FALSE)
}
\arguments{
\item{cluster}{list describing a cluster}

\item{material}{list}

\item{medium}{medium refractive index}

\item{Incidence}{angular directions of incident field}

\item{Axes}{incident field rotation axis}

\item{polarisation}{linear or circular polarisation}

\item{method}{linear system (ls), or order-of-scattering (oos)}

\item{cg}{logical, use conjugate gradient solver}

\item{born}{logical, use first Born approx as cg guess}

\item{Nsca}{number of quadrature points in calculation of csca}

\item{maxiter}{integer termination of conjugate gradient solver}

\item{tol}{double, tolerance of conjugate gradient solver}

\item{progress}{logical, display progress bar}
}
\value{
data.frame
}
\description{
dispersion spectrum
}
\details{
dispersion spectrum
}
\note{
The incident wavevector is along the z direction.
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda spectrum: \code{\link{spectrum_shell}}
}

