% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{spectrum_shell}
\alias{spectrum_shell}
\title{spectrum_shell}
\usage{
spectrum_shell(cluster, material, medium = 1.33, core = NULL,
  quadrature = c("gl", "qmc", "random", "cheap"), Nq = 100,
  iterative = FALSE, precision = 0.001, Qmax = 10000, dN = Nq,
  method = c("ls", "oos"), Nsca = 50, cg = FALSE, born = FALSE,
  maxiter = 30, tol = 1e-04, progress = FALSE, verbose = TRUE)
}
\arguments{
\item{cluster}{cluster (list)}

\item{material}{material}

\item{medium}{refractive index medium}

\item{core}{list}

\item{quadrature}{quadrature method, using either Gauss Legendre quadrature (default), Quasi Monte Carlo, random, or "cheap" (3 Axes)}

\item{Nq}{number of integration points}

\item{iterative}{logical, increase N until convergence (QMC only)}

\item{precision}{relative diff between two runs (QMC only)}

\item{Qmax}{maximum N if convergence not attained (QMC only)}

\item{dN}{iterative increase in N (QMC only)}

\item{method}{linear system (ls), or order-of-scattering (oos - not implemented)}

\item{Nsca}{quadrature points for scattering cross-section}

\item{cg}{logical use conjugate gradient solver}

\item{born}{logical first Born approx as cg guess}

\item{maxiter}{integer termination of conjugate gradient solver}

\item{tol}{double, tolerance of conjugate gradient solver}

\item{progress}{print progress lines}

\item{verbose}{display messages}
}
\value{
data.frame
}
\description{
Spectral simulation of a spherical shell of particles
}
\details{
Particles are arranged in a spherical shell around a central core (currently void).
A typical application will be to model the response of molecular layers.
}
\note{
Core is not yet implemented (will rely on Mie theory)
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda spectrum: \code{\link{spectrum_dispersion}}
}

