% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{spheroid_ar}
\alias{spheroid_ar}
\title{spheroid_ar}
\usage{
spheroid_ar(rv, h, type = c("prolate", "oblate"))
}
\description{
Spheroid described by effective radius and aspect ratio
}
\details{
Describe a spheroid by the aspect ratio and effective radius of an equi-volume sphere
V = 4/3 pi rv^3 = 4/3 pi a^2 c
c = h * a
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda utility: \code{\link{dye_coverage}}
}

