\name{table_10.3}
\docType{data}
\alias{table_10.3}
\title{Opinions About Birth Control and Premarital Sex Data Set of Table 10.3}
\description{Opinions about birth control and premarital sex data set of Table 10.3.}
\usage{data(table_10.3)}
\format{A data frame containing 16 observations (rows)
  and the following columns:
  \describe{
  \item{premar}{}
  \item{birth}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
