\name{table_12.3}
\docType{data}
\alias{table_12.3}
\title{Insomnia Data Set of Table 12.3}
\description{Insomnia data set of Table 12.3.}
\usage{data(table_12.3)}
\format{A data frame containing 478 observations (rows)
  and the following columns:
  \describe{
  \item{case}{}
  \item{treat}{}
  \item{occasion}{}
  \item{outcome}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
