\name{table_13.3}
\docType{data}
\alias{table_13.3}
\title{Attitudes About Abortion Data Set of Table 13.3, Incontigency Table Form}
\description{Attitudes about abortion data set of Table 13.3, incontigency table form.}
\usage{data(table_13.3)}
\format{A data frame containing 16 observations (rows)
  and the following columns:
  \describe{
  \item{gender}{}
  \item{poor}{}
  \item{single}{}
  \item{any}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
