\name{table_15.5}
\docType{data}
\alias{table_15.5}
\title{Election Data Set of Table 15.5 (1 = Dem, 0 = Rep)}
\description{Election data set of Table 15.5 (1 = Dem, 0 = Rep).}
\usage{data(table_15.5)}
\format{A data frame containing 51 observations (rows)
  and the following columns:
  \describe{
  \item{State}{}
  \item{e1}{}
  \item{e2}{}
  \item{e3}{}
  \item{e4}{}
  \item{e5}{}
  \item{e6}{}
  \item{e7}{}
  \item{e8}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
