\name{table_4.7}
\docType{data}
\alias{table_4.7}
\title{Teratology Study Data Set of Table 4.7}
\description{Teratology study data set of Table 4.7.}
\usage{data(table_4.7)}
\format{A data frame containing 58 observations (rows)
  and the following columns:
  \describe{
  \item{litter}{}
  \item{group}{}
  \item{n}{}
  \item{y}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
