\name{table_6.9}
\docType{data}
\alias{table_6.9}
\title{Clinical Trials Data Set of Table 6.9}
\description{Clinical trials data set of Table 6.9.}
\usage{data(table_6.9)}
\format{A data frame containing 32 observations (rows)
  and the following columns:
  \describe{
  \item{center}{}
  \item{treat}{}
  \item{response}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
