\name{table_8.5}
\docType{data}
\alias{table_8.5}
\title{Full Data Set for Table 8.5 on Happiness, Traumatic Events,  and Race}
\description{Full data set for Table 8.5 on happiness, traumatic events,  and race.}
\usage{data(table_8.5)}
\format{A data frame containing 97 observations (rows)
  and the following columns:
  \describe{
  \item{race}{}
  \item{trauma}{}
  \item{happy}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
