% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{visualise}
\alias{visualise}
\title{visualise}
\usage{
visualise(x, type, outfile = NULL, ...)
}
\arguments{
\item{x}{cluster}

\item{type}{type of visualisation (rgl or povray output)}

\item{outfile}{optional output file for the results}

\item{...}{additional arguments passed to the visualise method}
}
\description{
Visualise a cluster of particles
}
\details{
Helper function for rapid visualisation of cluster geometries.
}
\author{
baptiste Auguie
}

