% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povray.r
\name{visualise_povray}
\alias{visualise_povray}
\title{visualise_povray}
\usage{
visualise_povray(cl, outfile = "positions.pov")
}
\arguments{
\item{cl}{cluster}

\item{outfile}{output filename}
}
\value{
side-effect only
}
\description{
create include file for povray
}
\details{
writes a list of particles
}
\author{
baptiste Auguie
}

