% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizability.R
\name{alpha_kuwata}
\alias{Kuwata.A}
\alias{Kuwata.B}
\alias{alpha_kuwata}
\title{alpha_kuwata}
\usage{
alpha_kuwata(wavelength, epsilon, V, axis, L, medium = 1.33)
}
\arguments{
\item{wavelength}{wavelength}

\item{epsilon}{permittivity}

\item{V}{volume}

\item{axis}{semi-axis along incident field}

\item{L}{shape factor}

\item{medium}{refractive index}
}
\value{
polarizability
}
\description{
polarizability
}
\details{
prescription from Kuwata
}
\author{
baptiste Auguie
}
\references{
Kuwata et al. Resonant light scattering from metal nanoparticles: Practical analysis beyond Rayleigh approximation Appl. Phys. Lett. 83, 22 (2003)
}
\seealso{
Other user_level polarizability: \code{\link{depolarisation}}
}

