\name{alpha_kuwata}
\alias{alpha_kuwata}
\title{alpha_kuwata}
\usage{alpha_kuwata(lambda, epsilon, V, axis, L, epsilon.medium=1.33^2)
}
\description{polarizability}
\details{prescription from Kuwata}
\alias{alpha_kuwata}
\alias{Kuwata.A}
\alias{Kuwata.B}
\value{polarizability}
\author{baptiste Auguie}
\arguments{\item{lambda}{wavelength}
\item{epsilon}{permittivity}
\item{V}{volume}
\item{axis}{semi-axis along incident field}
\item{L}{shape factor}
\item{epsilon.medium}{medium permittivity}
}

