% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{array_factor}
\alias{array_factor}
\title{array factor}
\usage{
array_factor(wavelength, N, pitch)
}
\arguments{
\item{wavelength}{wavelength in nm}

\item{N}{half the number of dipoles along one side}

\item{pitch}{pitch in nm}
}
\value{
complex array factor
}
\description{
C++ calculation of the array factor
}
\details{
Brute-force numerical evaluation of the truncated 2D sum of dipole fields in a finite square array
}
\examples{
S <- array_factor(seq(400, 600),  10,  500)
str(S)
}
\author{
baptiste Auguie
}

