\docType{data}
\name{cd}
\alias{cd}
\title{Rcpp module: cd}
\description{
Exposes a calculation of orientation-averaged circular
dichroism within the coupled-dipole approximation.
}
\details{
\itemize{ \item{average_spectrum}{ Loop over wavelengths
and calculate the orientation averaging of the difference
in extinction, absorption, scattering for left/right
circularly polarised light} }
}
\examples{
show( cd )
}

