\name{cda-package}
\alias{cda-package}
\title{cda-package}
\description{Coupled dipole approximation}
\details{\tabular{ll}{
Package: \tab cda\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-15\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab false\cr
}

c++ implementation of the coupled dipole approximation for scattering calculations of interacting nanoantennas, specifically targetted at plasmonic nanoparticles}

\docType{package}
\author{baptiste Auguie \email{baptiste.auguie@gmail.com}}
\references{Draine BT. The discrete-dipole approximation and its application to interstellar graphite grains. Astrophysical Journal. 1988.

Schatz GC, Duyne RP. Discrete dipole approximation for calculating extinction and Raman intensities for small particles with arbitrary shapes. Journal of Chemical Physics. 1995.

Gunnarsson L, Zou S, Schatz GC, et al. Confined plasmons in nanofabricated single silver particle pairs: Experimental observations of strong interparticle interactions. Journal of Physical Chemistry B. 2005.

## Any one of the following references may (should) be used to cite and acknowledge this package.
A. Guerrero-Martinez, J. L. Alonso-Gomez, B. Auguie, M. Cid, and L. M.Liz-Marzan. From individual to collective chirality in metal nanoparticles. Nano Today, in press, 2011. doi:10.1016/j.nantod.2011.06.003 

A. Guerrero-Martinez, B. Auguie, J.L. Alonso-Gomez, Z. Dzolic, S. Gomez-Grana, M. Zinic, M.M. Cid, L.M. Liz-Marzan. Intense Optical Activity from three-Dimensional Chiral Ordering of Plasmonic Nanoantennas. Angew. Chem. Int. Ed.50 (2011)

B. Auguie, J.L. Alonso-Gomez, A. Guerrero-Martinez, L.M. Liz-Marzan. Fingers crossed: circular dichroism with a dimer of plasmonic nanorods. J. Phys. Chem. Lett. 2, (2011)

B. Auguie, W.L. Barnes. Diffractive coupling in gold nanoparticle arrays and the effect of disorder. Optics Letters (2009)

B. Auguie, W.L. Barnes. Collective resonances in gold nanoparticle arrays. Physical Review Letters (2008)}
\keyword{packagelibrary}

