\name{clust.equalangles}
\alias{clust.equalangles}
\title{clust.equalangles}
\usage{clust.equalangles(phi=0, theta=0, psi=0, N)
}
\description{clust.equalangles}
\details{generate a matrix of equal angles}
\value{matrix Nx3}
\author{baptiste Auguie}
\arguments{\item{phi}{phi}
\item{theta}{theta}
\item{psi}{psi}
\item{N}{N}
}

