\name{clust.equalsizes}
\alias{clust.equalsizes}
\title{clust.equalsizes}
\usage{
  clust.equalsizes(a, b, c, N)
}
\arguments{
  \item{a}{a}

  \item{b}{b}

  \item{c}{c}

  \item{N}{N}
}
\value{
  matrix Nx3
}
\description{
  clust.equalsizes
}
\details{
  generate a matrix of equal particle sizes
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level cluster:
  \code{\link{clust.equalangles}}, \code{\link{helix}},
  \code{\link{helix.zt}}, \code{\link{makeDimerCluster}},
  \code{\link{makeDimerDihedral}},
  \code{\link{makeHelixCluster}},
  \code{\link{makeHelixCluster.zt}},
  \code{\link{makeRodChain}},
  \code{\link{makeSpheresCluster}}
}

