% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_array}
\alias{cluster_array}
\title{cluster_array}
\usage{
cluster_array(N, pitch = 500, a = 50, b = 50, c = b)
}
\arguments{
\item{N}{number of particles}

\item{pitch}{center-to-center distance}

\item{a}{semi-axis along x}

\item{b}{semi-axis along y}

\item{c}{semi-axis along z}
}
\value{
list of class cluster with fields: positions, sizes,  angles
}
\description{
Square array of particles
}
\details{
A cluster describing a 2D square array of identical particles
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_ball}},
  \code{\link{cluster_chain}}, \code{\link{cluster_dimer}},
  \code{\link{cluster_helix}}, \code{\link{cluster_shell}},
  \code{\link{cluster_single}}
}

