% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_dimer}
\alias{cluster_dimer}
\title{cluster_dimer}
\usage{
cluster_dimer(d = 100, a = 35, b = 12, c = b, dihedral = pi/4,
  alpha1 = 0, alpha2 = 0)
}
\arguments{
\item{d}{center-to-center distance}

\item{a}{semi-axis along x}

\item{b}{semi-axis along y}

\item{c}{semi-axis along z}

\item{dihedral}{dihedral angle}

\item{alpha1}{angle first rod}

\item{alpha2}{angle second rod}
}
\value{
list of class cluster with fields: positions, sizes,  angles
}
\description{
A dimer of two particles
}
\details{
A cluster describing two particles, with dimer axis along z
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_array}},
  \code{\link{cluster_ball}}, \code{\link{cluster_chain}},
  \code{\link{cluster_helix}}, \code{\link{cluster_shell}},
  \code{\link{cluster_single}}
}

