% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_single}
\alias{cluster_single}
\title{cluster_single}
\usage{
cluster_single(a, b = a, c = b, phi = 0, theta = 0, psi = 0)
}
\arguments{
\item{a}{semi-axis along x}

\item{b}{semi-axis along y}

\item{c}{semi-axis along z}

\item{phi}{first Euler angle}

\item{theta}{second Euler angle}

\item{psi}{third Euler angle}
}
\value{
list of class cluster with fields: positions, sizes,  angles
}
\description{
Trivial cluster
}
\details{
A single particle cluster
}
\examples{
cl = cluster_single(10)
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_array}},
  \code{\link{cluster_ball}}, \code{\link{cluster_chain}},
  \code{\link{cluster_dimer}}, \code{\link{cluster_helix}},
  \code{\link{cluster_shell}}
}

