\name{curve.povray}
\alias{curve.povray}
\title{curve.povray}
\usage{curve.povray(positions, size=0.005, radius, out="positions.pov")
}
\description{create long string of small particles for povray input}
\details{writes a list of particles, as well as cylinder}
\value{side-effect only (note append=TRUE)}
\author{baptiste Auguie}
\arguments{\item{positions}{matrix of positions}
\item{size}{radius of particles}
\item{radius}{radius of inner cylinder}
\item{out}{output filename}
}

