\docType{data}
\name{dispersion}
\alias{dispersion}
\title{Rcpp module: dispersion}
\description{
Exposes C++ calculation of scattering and absorption of
dipolar particles by linearly polarised light in fixed
orientation.
}
\details{
\itemize{ \item{dispersion}{Returns absorption, scattering
and extinction cross-sections for two orthogonal
polarisations at multiple angles of incidence, fixed
wavelength (subroutine)} \item{dispersion_spectrum}{Returns
absorption, scattering and extinction spectra for two
orthogonal polarisations at multiple angles of incidence} }
}
\examples{
show( dispersion )
}

