\name{dispersion_spectrum}
\alias{dispersion_spectrum}
\title{dispersion_spectrum}
\usage{
  dispersion_spectrum(cluster, angles, material, n = 1.33,
    invert = FALSE, progress = FALSE)
}
\arguments{
  \item{cluster}{list describing a cluster}

  \item{angles}{matrix of incident Euler angles}

  \item{material}{list}

  \item{n}{medium refractive index}

  \item{invert}{logical}

  \item{progress}{logical, display progress bar}
}
\value{
  data.frame
}
\description{
  dispersion spectrum
}
\details{
  dispersion spectrum
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level linear:
  \code{\link{linear_extinction_spectrum}}
}

