\name{dispersion_spectrum}
\alias{dispersion_spectrum}
\title{dispersion_spectrum}
\usage{
dispersion_spectrum(cluster, material, medium = 1.33, angles = 0,
  axes = "z", polarisation = c("linear", "circular"), progress = FALSE)
}
\arguments{
  \item{cluster}{list describing a cluster}

  \item{material}{list}

  \item{medium}{medium refractive index}

  \item{angles}{of incident field in radians}

  \item{axes}{of incident field rotation character vector
  from ('x', 'y', 'z')}

  \item{polarisation}{linear or circular polarisation}

  \item{progress}{logical, display progress bar}
}
\value{
data.frame
}
\description{
dispersion spectrum
}
\details{
dispersion spectrum
}
\author{
baptiste Auguie
}

