% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dye_coverage}
\alias{dye_coverage}
\title{dye_coverage}
\usage{
dye_coverage(rho, R)
}
\arguments{
\item{rho}{surface demsity}

\item{R}{radius}
}
\description{
dye_coverage
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda utility: \code{\link{equal_angles}},
  \code{\link{equal_sizes}}, \code{\link{spheroid_ar}}
}

