\name{equal_angles}
\alias{equal_angles}
\title{equal_angles}
\usage{
equal_angles(phi = 0, theta = 0, psi = 0, N)
}
\arguments{
  \item{phi}{phi}

  \item{theta}{theta}

  \item{psi}{psi}

  \item{N}{N}
}
\value{
matrix Nx3
}
\description{
equal_angles
}
\details{
generate a matrix of equal angles
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_chain}},
\code{\link{cluster_dimer}},
\code{\link{cluster_dimer_end}},
\code{\link{cluster_helix}}, \code{\link{equal_sizes}},
\code{\link{helix}}
}

