\name{equal_sizes}
\alias{equal_sizes}
\title{equal_sizes}
\usage{
  equal_sizes(a, b, c, N)
}
\arguments{
  \item{a}{a}

  \item{b}{b}

  \item{c}{c}

  \item{N}{N}
}
\value{
  matrix Nx3
}
\description{
  equal_sizes
}
\details{
  generate a matrix of equal particle sizes
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level cluster: \code{\link{cluster_chain}},
  \code{\link{cluster_dimer}}, \code{\link{cluster_helix}},
  \code{\link{equal_angles}}, \code{\link{helix}}
}

