\docType{data}
\name{G0}
\alias{G0}
\alias{gfun}
\title{Lattice sum}
\format{\describe{
   \item{\code{wavelength}}{a numeric vector}
   \item{\code{Qx}}{a numeric vector}
   \item{\code{Gxx}}{a complex vector}
 }}
\description{
Converged lattice sum G0 for an infinite 2D array of
dipoles at normal incidence
}
\details{
The calculation was made using code from Prof. J. G. de
Abajo (CSIC, Spain)
}
\examples{
data(G0)
\dontrun{demo(lattice_sum)}
}

