\name{La}
\alias{La}
\title{La}
\usage{La(a=0.05, b=a, c=a)
}
\description{Shape factor for an ellipsoid}
\details{calculates the shape factor for a general ellipsoid}
\value{shape factor along a}
\author{baptiste Auguie}
\arguments{\item{a}{semi-axis in um}
\item{b}{semi-axis in um}
\item{c}{semi-axis in um}
}

