\name{La}
\alias{La}
\title{La}
\usage{
  La(a = 0.05, b = a, c = a)
}
\arguments{
  \item{a}{semi-axis in um}

  \item{b}{semi-axis in um}

  \item{c}{semi-axis in um}
}
\value{
  shape factor along a
}
\description{
  Shape factor for an ellipsoid
}
\details{
  calculates the shape factor for a general ellipsoid
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level polarizability:
  \code{\link{alpha_kuwata}}, \code{\link{make.invalpha}},
  \code{\link{polarizability.ellipsoid}}
}

