\name{La}
\alias{La}
\title{La}
\usage{
  La(a = 50, b = a, c = a)
}
\arguments{
  \item{a}{semi-axis in nm}

  \item{b}{semi-axis in nm}

  \item{c}{semi-axis in nm}
}
\value{
  shape factor along a
}
\description{
  Shape factor for an ellipsoid
}
\details{
  calculates the shape factor for a general ellipsoid
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level polarizability:
  \code{\link{alpha_kuwata}},
  \code{\link{inverse_polarizability}},
  \code{\link{polarizability_ellipsoid}}
}

