\name{linear_extinction_spectrum}
\alias{linear_extinction_spectrum}
\title{extinction_spectrum}
\usage{
  linear_extinction_spectrum(cluster, material, n = 1.33,
    progress = FALSE)
}
\arguments{
  \item{cluster}{list describing a cluster}

  \item{material}{list}

  \item{n}{medium refractive index}

  \item{progress}{logical, display progress bar}
}
\value{
  data.frame
}
\description{
  extinction spectrum
}
\details{
  extinction spectrum
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level linear:
  \code{\link{dispersion_spectrum}}
}

