\name{makeHelixCluster.zt}
\alias{makeHelixCluster.zt}
\title{makeHelixCluster.zt}
\usage{
  makeHelixCluster.zt(z = runif(10, -0.5, 0.5),
    theta = runif(10, 0, 2 * pi), R0 = 0.5, pitch = 6,
    delta0 = 0, right = TRUE, a = 0.035, b = 0.012,
    c = 0.012)
}
\arguments{
  \item{z}{z}

  \item{theta}{theta}

  \item{R0}{R0}

  \item{pitch}{pitch}

  \item{delta0}{delta0}

  \item{right}{right}

  \item{a}{a}

  \item{b}{b}

  \item{c}{c}
}
\value{
  cluster
}
\description{
  helical cluster z theta
}
\details{
  helical cluster z theta
}
\author{
  baptiste Auguie
}
\seealso{
  Other user_level cluster:
  \code{\link{clust.equalangles}},
  \code{\link{clust.equalsizes}}, \code{\link{helix}},
  \code{\link{helix.zt}}, \code{\link{makeDimerCluster}},
  \code{\link{makeDimerDihedral}},
  \code{\link{makeHelixCluster}},
  \code{\link{makeRodChain}},
  \code{\link{makeSpheresCluster}}
}

