% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{spectrum_shell}
\alias{spectrum_shell}
\title{spectrum_shell}
\usage{
spectrum_shell(cluster, wavelength, fun = polarizability_dye, ...,
  medium = 1.33, core = TRUE, Nq = 100, quadrature = c("cheap", "qmc",
  "gl"), method = c("ls", "oos"), iterative = FALSE, precision = 0.001,
  Qmax = 10000, dN = Nq, cg = TRUE, born = TRUE, nmax = 10,
  tol = 0.01, full = TRUE, progress = FALSE, verbose = TRUE,
  result.matrix = FALSE)
}
\arguments{
\item{cluster}{list describing a cluster}

\item{medium}{medium refractive index}

\item{cg}{logical, use conjugate gradient solver}

\item{born}{logical, use first Born approx as cg guess}

\item{nmax}{integer termination of conjugate gradient solver}

\item{tol}{double, tolerance of conjugate gradient solver}

\item{progress}{logical, display progress bar}

\item{material}{list}

\item{angles}{of incident field in radians}

\item{axes}{of incident field rotation character vector from ('x', 'y', 'z')}

\item{polarisation}{linear or circular polarisation}

\item{method:}{linear system (ls), or order-of-scattering (oos)}
}
\value{
data.frame
}
\description{
shell spectrum
}
\details{
shell spectrum
}
\note{
The incident wavevector is along the z direction.
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cda spectrum: \code{\link{spectrum_dispersion}}
}

